+++ lua/glayout/cl_init.lua
@@ -4,32 +4,55 @@
 function DrawHUD()
 
     -- Create our new grid
-    local grid = Grid:Create()
+    local grid  = Grid:Create(
+    {
+        -- Position
+        pos         =
+        {
+            x       = 0,
+            y       = ScrH() - ( ScrH() / 10 )
+        },
+        -- Width
+        size        =
+        {
+            width   = ScrW(),
+        },
+        -- Margins
+        margin      =
+        {
+            right   = 20,
+            bottom  = 20,
+            left    = 20
+        },
+    })
 
-        -- Declare Sizing & Positioning of Grid
-        grid:SetPos( 0, ScrH() - ( ScrH() / 10 ) )
-
-        -- Set Grid Properties
-        grid:SetWidth( ScrW() )
-        grid:SetMargin( { 0, 20, 20, 20 } )
-
-        -- Initialize Our Grid
-        grid:Init()
+    -- Initialize Our Grid
+    grid:Init()
 
 
     -- Create Columns
     local col = {}
 
+    -- Set Column Styles
+    local colStyles =
+    {
+        span        = 6,
+        size        =
+        {
+            height      = ScrH() / 10,
+        },
+    }
+
 
     -- Create First Column
-    col[1] = grid:CreateCol( 6 )
+    col[1] = grid:CreateCol( colStyles )
 
-    -- Set Column Properties
-    col[1]:SetHeight( ScrH() / 10 )
-    col[1]:SetMarginRight( 10 )
+        -- Set Margin-Right for this column only
+        col[1]:SetMarginRight( 10 )
 
     -- Initialize Our Column
     col[1]:Init()
+
 
     -- Stores player in local variable
     local ply = LocalPlayer()
@@ -46,11 +69,10 @@ function DrawHUD()
 
 
     -- Create Second Column
-    col[2] = grid:CreateCol( 6 )
+    col[2] = grid:CreateCol( colStyles )
 
-    -- Set Column Properties
-    col[2]:SetHeight( ScrH() / 10 )
-    col[2]:SetMarginLeft( 10 )
+        -- Set Margin-Left for this column only
+        col[2]:SetMarginLeft( 10 )
 
     -- Initialize Our Column
     col[2]:Init()